+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2020-03-06
 ### Added
 - CSS handle when in the `modal-actions` block when you reach the bottom of the content in scroll.
 - Prop `iconCloseSize` to the `modal-header` block.